Add files via upload
Given a stack of integers of size N, check whether numbers in the stack are pairwise consecutive or not. The pairs can be increasing or decreasing, and if the stack has an odd number of elements, the element at the top is left out of a pair. The code should retain the original stack content.

Input:
3
6
1 2 3 4 5 6
5
1 2 4 5 7
5
1 5 3 9 7
Output:
Yes
Yes
No